+++ src/dom/mod.rs
@@ -2,7 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-pub use self::doc::Document;
+pub use self::document::Document;
 pub use self::element_type::ElementType;
 pub use self::iterators::*;
 pub use self::node::Node;
@@ -14,7 +14,7 @@ pub type TagNameRef<'a> = NameRef<'a, ElementId>;
 /// Type alias for `Name<ElementId>`.
 pub type TagName = Name<ElementId>;
 
-mod doc;
+mod document;
 mod element_type;
 mod iterators;
 mod node;