+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.4.0] - 2017-03-15
 ### Added
 - `Node::make_copy` and `Node::make_deep_copy`.
 - `Error::InvalidEncoding` and `Error::Utf8Error`.
@@ -179,7 +181,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/RazrFalcon/libsvgdom/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/RazrFalcon/libsvgdom/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/RazrFalcon/libsvgdom/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/RazrFalcon/libsvgdom/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/RazrFalcon/libsvgdom/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/RazrFalcon/libsvgdom/compare/v0.1.0...v0.2.0